+++ frontend/src/modules/toggleBtn.js
@@ -3,7 +3,7 @@ export function toggleBtn() {
     if (e.target.type === 'textarea') {
       const textarea = e.target;
       const addBtn = textarea.nextElementSibling.firstElementChild;
-      if (textarea.value.length !== 0) {
+      if (textarea.value.length !== '') {
         addBtn.removeAttribute('disabled');
       } else {
         addBtn.setAttribute('disabled', 'true');
